requirements.txt
pyswisseph==2.10.3.1
python-dotenv==1.0.0
python_telegram_bot==13.11
